# BACKEND/collection/openapi.yaml
openapi: 3.0.1
info:
  title: LID - Life Event Distribution API
  description: API REST complète pour la gestion d'une plateforme d'e-commerce de
    distribution d'événements
  contact:
    name: LID Team
    email: support@lid.com
  license:
    name: Apache 2.0
  version: 1.0.0
servers:
- url: http://localhost:9000
  description: Development Server
- url: https://api.lid.com
  description: Production Server
...
paths:
...
  /api/categories/{id}:
    get:
      tags:
      - category-controller
      operationId: getCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryDto"
    put:
      tags:
      - category-controller
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryDto"
    delete:
      tags:
      - category-controller
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
...
  /api/categories:
    get:
      tags:
      - category-controller
      operationId: getAllCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryDto"
    post:
      tags:
      - category-controller
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryDto"
...
  /api/categories/name/{name}:
    get:
      tags:
      - category-controller
      operationId: getCategoryByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryDto"
...
  /api/articles/category/{categoryId}:
    get:
      tags:
      - Articles
      summary: Récupérer les articles d'une catégorie
      description: Récupère tous les articles d'une catégorie spécifique
      operationId: getByCategory
      parameters:
      - name: categoryId
        in: path
        description: ID de la catégorie
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Articles de la catégorie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDto"
...
components:
  schemas:
...
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        orderIdx:
          type: integer
          format: int32
    ArticleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        img:
          type: string
        sku:
          type: string
        ean:
          type: string
        vat:
          type: number
          format: float
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDto"
...
  securitySchemes:
    Bearer Token:
      type: http
      description: Enter your JWT token
      in: header
      scheme: bearer
      bearerFormat: JWT